requirements.txt
celery==5.4.0
fastapi==0.115.3
flower==2.0.1
httpcore==1.0.7
httpx==0.26.0
ollama-python==0.1.2
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.4
python-dotenv==1.0.1
redis==5.2.0
regex==2024.9.11
requests==2.31.0
stack-data==0.6.3
typing_extensions==4.12.2
uvicorn==0.32.0

